+++ packages/nimble-components/src/menu-button/tests/menu-button.stories.ts
@@ -1,4 +1,5 @@
 import { html, when } from '@microsoft/fast-element';
+import { withActions } from '@storybook/addon-actions/decorator';
 import type { Meta, StoryObj } from '@storybook/html';
 import { createUserSelectedThemeStory } from '../../utilities/tests/storybook';
 import { ButtonAppearance, MenuButtonPosition } from '../types';
@@ -30,6 +31,7 @@ This icon will be hidden when \`contentHidden\` is set to \`true\`
 const metadata: Meta<MenuButtonArgs> = {
     title: 'Menu Button',
     tags: ['autodocs'],
+    decorators: [withActions],
     parameters: {
         docs: {
             description: {