+++ packages/nimble-components/src/table/components/row/template.ts
@@ -10,9 +10,8 @@ export const template = html<TableRow>`
             ${when(x => !x.column.columnHidden, html<ColumnState, TableRow>`
                 <${tableCellTag}
                     class="cell"
-                    :cellTemplate="${x => x.column.cellTemplate}"
-                    :cellStyles="${x => x.column.cellStyles}"
                     :cellState="${x => x.cellState}"
+                    :cellViewTemplate="${x => x.column.currentCellViewTemplate}"
                     ?has-action-menu="${x => !!x.column.actionMenuSlot}"
                     action-menu-label="${x => x.column.actionMenuLabel}"
                     @cell-action-menu-beforetoggle="${(x, c) => c.parent.onCellActionMenuBeforeToggle(c.event as CustomEvent<MenuButtonToggleEventDetail>, x.column)}"