+++ packages/nimble-components/src/utilities/tests/storybook-declarations.d.ts
@@ -1,0 +1,3 @@
+// This file is a workaround to allow Nimble to import types that Storybook doesn't export correctly
+// https://github.com/storybookjs/storybook/issues/21887#issuecomment-1495427769
+declare module '@storybook/addon-actions/decorator';